Refactoring : 파일 분할
