+++ app/src/main/java/com/example/byeoldori/skymap/components/RealtimeItemCard.kt
@@ -1,0 +1,70 @@
+package com.example.byeoldori.skymap.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.byeoldori.skymap.viewmodel.ObjectItem
+import com.example.byeoldori.ui.theme.Purple800
+import com.example.byeoldori.ui.theme.TextHighlight
+
+
+@Composable
+fun RealtimeItemCard(
+    item: ObjectItem,
+    modifier: Modifier = Modifier
+) {
+    // 🔹 공백 3개 기준 분할
+    val splitValues = remember(item.value) {
+        if (item.value.contains("   ")) {
+            item.value.split("   ").map { it.trim() }
+        } else {
+            listOf(item.value)
+        }
+    }
+
+    Card(
+        shape = RoundedCornerShape(12.dp),
+        colors = CardDefaults.cardColors(containerColor = Purple800),
+        modifier = modifier.padding(4.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(12.dp)
+                .fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(4.dp)
+        ) {
+            // Key
+            Text(
+                text = item.key,
+                color = TextHighlight,
+                fontWeight = FontWeight.Bold,
+                fontSize = 24.sp,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+
+            // Value (세로 배치)
+            splitValues.forEach { part ->
+                Text(
+                    text = part,
+                    color = TextHighlight,
+                    fontSize = 16.sp,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+            }
+        }
+    }
+}
